Get some google services under control
It has been observed that google services create some unfortunate wakelocks that pretty much prevent any device from getting to deep sleep. Network Locations Provider (nlp) wakelock is apparently used for keeping track of the device's location (low accuracy), but renewing the low accuracy location every 5 seconds is WAY too frequent. Frankly, when the device is asleep, there is really not much use in keeping a frequently updated low accuracy location. If you need to keep a log of your location, then create another process to frequently wake up the device and renew the location. This will make sure that wakelock is never created, thus nlp can only obtain the device's location when the device is awake as a result of some other process (like the screen being on). We also block SystemUpdate from obtaining a wakelock -- it perpetually tries to pull in a system update that does not exist.

We also block nlp from creating alarms to wake the device.

Right now, all these blocks are hardcoded. This will be enhanced with a database to store the list of items to be blocked.